agent_service/pyproject.toml
[project]
name = "agent-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[all]>=0.115.12",
    "msal>=1.32.3",
    "pydantic-settings>=2.9.1",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.5.0",
    "requests>=2.32.4",
    "uvicorn[standard]>=0.34.3",
]
